expose bind_server, and deduplicate plain / tls
This exposes a top-level bind_server and de-duplicates plain and tls
serving to use it. The underlying goal is to allow clients to use other
TLS implementations if they'd like.

There's a couple behavior changes this causes:

- This adds one extra function call when accepting plain connections,
  but that seems like it shouldn't amount to anything.
- This removes the panics in the TLS code path. Those could be triggered
  with an invalid handshake (e.g. a client connecting over HTTP to a
  HTTPS server), so it's desirable to take them away.
