+++ gotham/Cargo.toml
@@ -44,6 +44,7 @@ http = "0.1"
 httpdate = "0.3"
 failure = "0.1"
 tokio-rustls = "0.9"
+tokio-io = "0.1"
 
 [dev-dependencies]
 gotham_derive = "0.5.0-dev"
